[util] Moving symbolic_types to util
