+++ src/clang-c-frontend/clang_c_convert.cpp
@@ -16,7 +16,6 @@ CC_DIAGNOSTIC_IGNORE_LLVM_CHECKS()
 CC_DIAGNOSTIC_POP()
 
 #include <ac_config.h>
-#include <clang-c-frontend/symbolic_types.h>
 #include <clang-c-frontend/clang_c_convert.h>
 #include <clang-c-frontend/typecast.h>
 #include <util/arith_tools.h>
@@ -28,6 +27,8 @@ CC_DIAGNOSTIC_POP()
 #include <util/mp_arith.h>
 #include <util/std_code.h>
 #include <util/std_expr.h>
+#include <util/symbolic_types.h>
+
 #include <boost/algorithm/string/replace.hpp>
 
 clang_c_convertert::clang_c_convertert(
