+++ src/util/CMakeLists.txt
@@ -24,7 +24,7 @@ add_library(util_esbmc xml_irep.cpp xml.cpp
         string_container.cpp options.cpp c_misc.cpp
         simplify_expr.cpp dstring.cpp simplify_expr2.cpp
         simplify_utils.cpp string2array.cpp array2string.cpp time_stopping.cpp symbol.cpp
-        symbol_generator.cpp
+        symbol_generator.cpp symbolic_types.cpp
         type_eq.cpp guard.cpp array_name.cpp union_find.cpp
         std_types.cpp std_code.cpp format_constant.cpp
         irep_serialization.cpp symbol_serialization.cpp fixedbv.cpp
