+++ src/util/symbolic_types.cpp
