+++ src/util/symbolic_types.h
